requirements.txt
uvicorn[standard]>=0.27.0.post1
fastapi>=0.109.2
python-multipart>=0.0.7
Jinja2>=3.1.3
pydantic>=2.6.1
PyYAML>=6.0.1
SPARQLWrapper>=2.0.0
pandas>=2.2.1